commit #72: code optimization

+++ server/src/controller/admin.js
@@ -61,7 +61,7 @@ const getBookingByMultiInputs = async (req, res) => {
     
         const query = "SELECT b.id, b.nb_adults, b.nb_children, b.price_total_booking, b.payment_type, b.status, b.date_created, u.first_name, u.last_name, u.email, u.tel, p.reference, p.departure_date, p.return_date, p.duration, d.name, d.country, d.departure_place FROM bookings AS b JOIN users AS u ON u.id = b.user_id JOIN packs AS p ON p.id = b.pack_id JOIN destinations AS d ON d.id = p.destination_id WHERE"+queryEnding;
 
-        const [datas] = await Query.findByArrayDatasNTables(query, bodyData);
+        const [datas] = await Query.queryByArrayNTables(query, bodyData);
 
         res.status(200).json({datas});
     } catch (err) {
@@ -92,7 +92,7 @@ const getDestinationByMultiInputs = async (req, res) => {
     
         const query = "SELECT d.reference, d.name, d.country, d.continent, d.overview, d.departure_place, d.date_created FROM destinations AS d WHERE"+queryEnding;
 
-        const [datas] = await Query.findByArrayDatasNTables(query, bodyData);
+        const [datas] = await Query.queryByArrayNTables(query, bodyData);
 
         res.status(200).json({datas});
     } catch (err) {
@@ -104,7 +104,7 @@ const getDestinationByMultiInputs = async (req, res) => {
 const getLodgingByMultiInputs = async (req, res) => {
     try {
         const query = "SELECT l.* FROM lodgings AS l WHERE l.name = ? LIMIT 1";
-        const [datas] = await Query.findByValueNTables(query, req.body.name);
+        const [datas] = await Query.queryByValueNTables(query, req.body.name);
         res.status(200).json({datas});
     } catch (err) {
         throw Error(err);
@@ -112,44 +112,61 @@ const getLodgingByMultiInputs = async (req, res) => {
 } 
 /* ----------------------- UPDATE ---------------------- */
 const modifyLodging = async (req, res) => {
-    try {
-        let msg = "";
-        //console.log(req.body);
-        const queryUser = "UPDATE lodgings SET name = ?, type = ?, overview = ?, facilities = ?, rooms = ?, food_drink = ?, meal_plans = ?, entertainment = ?, children = ?, tripadvisor = ?, coordinates = ? WHERE id = ?";
-        //const queryUser = "UPDATE lodgings SET name = 'test-changé', type = 'a1', overview = 'a1', facilities = 'a1', rooms = 'a1', food_drink = 'a1', meal_plans = 'a1', entertainment = 'a1', children = 'a1', tripadvisor = 'a1', coordinates = 'a1' WHERE id = 8";
-        await Query.update(queryUser, req.body);
-        msg = "hébérgement modifié";
-        res.status(201).json({ msg });
-    } catch (err) {
-        throw Error(err)
-    }
+    const form = formidable({
+        uploadDir: `public/img/lodgings`,
+        keepExtensions: true,
+        allowEmptyFiles: false,
+        multiples: true,
+    });
+
+    form.parse(req, async (err, fields, files) => {
+        try {
+            let msg = "";
+            const lodging = {};
+    
+            for (const key in fields){
+                lodging[key] = fields[key][0]; 
+            }
+            const id = lodging.id;
+            delete lodging.id;
+            lodging.url_initial_image = files.file[0].newFilename;
+            lodging.id = id;
+            console.log(lodging);
+            const queryUser = "UPDATE lodgings SET name = ?, type = ?, overview = ?, facilities = ?, rooms = ?, food_drink = ?, meal_plans = ?, entertainment = ?, children = ?, tripadvisor = ?, coordinates = ?, url_initial_image = ? WHERE id = ?";
+            await Query.queryByObject(queryUser, lodging);
+
+            msg = "hébérgement modifié";
+            res.status(200).json({ msg });
+        } catch (err) {
+            throw Error(err)
+        }
+    });
 }
 
 /* ----------------------- CREATE ----------------------- */
 const createLodging = async (req, res) => {
-    try {
-        //fs.mkdirSync(__dirname + `../../../public/img/lodgings`);
-        const form = formidable({
-            //uploadDir: __dirname + `../../../public/img/lodgings`,
-            uploadDir: `public/img/lodgings`,
-            keepExtensions: true,
-            allowEmptyFiles: false,
-            multiples: true,
-        });
+    //fs.mkdirSync(__dirname + `../../../public/img/lodgings`);
+    const form = formidable({
+        //uploadDir: __dirname + `../../../public/img/lodgings`,
+        uploadDir: `public/img/lodgings`,
+        keepExtensions: true,
+        allowEmptyFiles: false,
+        multiples: true,
+    });
 
-        form.parse(req, async (err, fields, files) => {
+    form.parse(req, async (err, fields, files) => {
+        try {
             let msg = "";
             const newLodging = {};
-
+    
             for (const key in fields){
                 newLodging[key] = fields[key][0]; 
             }
-
+    
             // vérifier si un hébérgement avec le même nom existe déjà :
             const queryCheck = "SELECT name FROM lodgings WHERE name = ?";
-            const [results] = await Query.findByValue(queryCheck, newLodging.name);
-
-            //console.log(results);
+            const [results] = await Query.queryByValue(queryCheck, newLodging.name);
+    
             if (results.length) {
                 msg = "Un hébérgement avec ce nom existe déjà !";
                 res.status(409).json({ msg });
@@ -157,12 +174,12 @@ const createLodging = async (req, res) => {
                 newLodging.url_initial_image = files.file[0].newFilename;
                 
                 const queryInsertLodg = "INSERT INTO lodgings (name, type, overview, facilities, rooms, food_drink, meal_plans, entertainment, children, tripadvisor, coordinates, url_initial_image) VALUES (?,?,?,?,?,?,?,?,?,?,?,?)";
-                const resultLodg = await Query.write(queryInsertLodg, newLodging);
-
+                const resultLodg = await Query.queryByObject(queryInsertLodg, newLodging);
+    
                 // on a besoin de récupérer l'ID de l'hébergement créé pour l'utiliser comme clé étrangère :
                 const queryGetID = "SELECT id FROM lodgings ORDER BY id DESC LIMIT 1";
                 const [lodgID] = await Query.find(queryGetID);
-
+    
                 // pour stocker les urls des images pour le slideshow :
                 const urlArray = [];
                 const images = Object.values(files);
@@ -176,22 +193,22 @@ const createLodging = async (req, res) => {
                 // on en aura pas besoin de la 1ere image, car elle sert uniquement pour image initiale :
                 urlArray.shift();
                 console.log(urlArray);
-
+    
                 async function writeImage(imgData){
                     const queryInsertImages = "INSERT INTO lodgings_images (url_image, lodging_id) VALUES (?,?)";
-                    await Query.write(queryInsertImages, imgData);
+                    await Query.queryByObject(queryInsertImages, imgData);
                 }
                 urlArray.forEach((el) => {
                     writeImage(el);
                 });
-
+    
                 msg = "Merci ! L'hébérgement a été créé dans la BDD.";
                 res.status(200).json({ msg });
             }
-        });
-    } catch (err) {
-        throw Error(err)
-    }
+        } catch (err) {
+            throw Error(err)
+        }
+    });
 }
 
 const createDestination = async (req, res) => {
@@ -214,7 +231,7 @@ const createDestination = async (req, res) => {
 
             // vérifier si une destination avec le même nom existe déjà :
             const queryCheck = "SELECT name FROM destinations WHERE name = ?";
-            const [results] = await Query.findByValue(queryCheck, queryCheck, newDestination.name);
+            const [results] = await Query.queryByValue(queryCheck, queryCheck, newDestination.name);
 
             if (results.length) {
                 msg = "Une destination avec ce nom existe déjà !";
@@ -223,7 +240,7 @@ const createDestination = async (req, res) => {
                 newDestination.url_initial_image = files.file[0].newFilename;
 
                 const queryInsertDest = "INSERT INTO destinations (reference, name, country, continent, overview, departure_place, date_created, lodging_id, url_initial_image) VALUES (?,?,?,?,?,?,CURRENT_TIMESTAMP(),?,?)";
-                const resultDest = await Query.write(queryInsertDest, newDestination);
+                const resultDest = await Query.queryByObject(queryInsertDest, newDestination);
 
                 // on a besoin de récupérer l'ID de la destination créée pour l'utiliser comme clé étrangère :
                 const queryGetID = "SELECT id FROM destinations ORDER BY id DESC LIMIT 1";
@@ -245,7 +262,7 @@ const createDestination = async (req, res) => {
 
                 async function writeImage(imgData){
                     const queryInsertImages = "INSERT INTO destinations_images (url_image, destination_id) VALUES (?,?)";
-                    await Query.write(queryInsertImages, imgData);
+                    await Query.queryByObject(queryInsertImages, imgData);
                 }
                 urlArray.forEach((el) => {
                     writeImage(el);
@@ -280,7 +297,7 @@ const createPack = async (req, res) => {
             console.log(newPack);
 
             const queryInsert = "INSERT INTO packs (reference, departure_date, return_date, duration, price_adults, price_children, discount, places_total, places_left, destination_id) VALUES (?,?,?,?,?,?,?,?,?,?)";
-            await Query.write(queryInsert, newPack);
+            await Query.queryByObject(queryInsert, newPack);
     
             msg = "Merci ! Le pack a été créé dans la BDD.";
             res.status(200).json({ msg });
@@ -296,7 +313,7 @@ const createActivity= async (req, res) => {
 
         // on vérifie si une activité avec le même nom existe déjà :
         const queryCheck = "SELECT id FROM activities WHERE name = ?";
-        const [results] = await Query.findByValue(queryCheck, req.body.nameAct);
+        const [results] = await Query.queryByValue(queryCheck, req.body.nameAct);
 
         if (results.length) {
             msg = "Un activité avec ce nome existe déjà !";
@@ -314,15 +331,15 @@ const createActivity= async (req, res) => {
                 placesLeft: req.body.placesLeft
             };
             const queryInsertAct = "INSERT INTO activities (name, type, overview, price_adults, price_children, places_total, places_left) VALUES (?,?,?,?,?,?,?)";
-            await Query.write(queryInsertAct, datas);
+            await Query.queryByObject(queryInsertAct, datas);
             // récupérer l'ID de l'activité insérée 
             const queryGetID = "SELECT id FROM activities WHERE name = ?";
-            const [activityData] = await Query.findByValue(queryGetID, req.body.nameAct);
+            const [activityData] = await Query.queryByValue(queryGetID, req.body.nameAct);
             console.log(activityData[0].id);
             // insérer l'activité (via son ID) dans le tableau 'destinations_activities' afin d'associer l'activité à la destination
             const dataArray = [activityData[0].id, req.body.destinationID];
             const queryInsertDestAct = "INSERT INTO destinations_activities (activity_id, destination_id) VALUES (?,?)";
-            await Query.writeWithArray(queryInsertDestAct, dataArray);
+            await Query.queryByArray(queryInsertDestAct, dataArray);
             // si tout est OK :
             msg = "Merci ! L'activité' a été créée dans la BDD.";
             res.status(200).json({ msg });
@@ -337,11 +354,26 @@ const createActivity= async (req, res) => {
 const deleteLodging = async (req, res) => {
     try {
         let msg = "";
-        //console.log(req.body.id);
-        const queryUser = "DELETE FROM lodgings WHERE id = ?";
-        await Query.delete(queryUser, req.body.id);
-        msg = "hébérgement supprimé";
-        res.status(201).json({ msg });
+        // vérifier si la destination est déjà supprimée :
+        const queryCheck = "SELECT id FROM destinations WHERE lodging_id = ?";
+        const [results] = await Query.queryByValue(queryCheck, req.body.id);
+        console.log(results);
+        console.log(results.length);
+        if (results.length) {
+            console.log("trouvée");
+            msg = "Veuillez supprimer d'abord la destination liée à cet hébergement.";
+            res.status(409).json({ msg });
+        } else if (!results.length){
+            console.log("PAS trouvée - OK");
+            // supprimer les images d'abord
+            const queryLodgImg = "DELETE FROM lodgings_images WHERE lodging_id = ?";
+            await Query.queryByValue(queryLodgImg, req.body.id);
+            // supprimer l'hébérgement
+            const queryLodg = "DELETE FROM lodgings WHERE id = ?";
+            await Query.queryByValue(queryLodg, req.body.id);
+            msg = "hébérgement supprimé";
+            res.status(200).json({ msg });
+        }
     } catch (err) {
         throw Error(err)
     }
